requirements.txt
streamlit
newspaper3k
lxml_html_clean